Move new_working_dir to the tests.helpers.io module
