+++ python/tests/helpers/io.py
@@ -8,8 +8,6 @@ import os
 from contextlib import contextmanager
 from tempfile import NamedTemporaryFile, TemporaryDirectory
 
-from aac.util import new_working_dir
-
 
 def clear_directory(dirspec: str, *file_list: list[str]) -> None:
     """Clear the specified directory of any file contents.
@@ -56,3 +54,32 @@ def temporary_test_file(content: str, **extra_file_attrs):
             file.seek(0)
 
             yield file
+
+
+@contextmanager
+def new_working_dir(directory):
+    """Change directories to execute some code, then change back.
+
+    Args:
+        directory: The new working directory to switch to.
+
+    Returns:
+        The new working directory.
+
+    Example Usage:
+        from os import getcwd
+        from tempfile import TemporaryDirectory
+
+        print(getcwd())
+        with TemporaryDirectory() as tmpdir, new_working_dir(tmpdir):
+            print(getcwd())
+        print(getcwd())
+    """
+    current_dir = os.getcwd()
+    try:
+        os.chdir(directory)
+        yield os.getcwd()
+        os.chdir(current_dir)
+    except Exception as e:
+        os.chdir(current_dir)
+        raise e
